Rearrange code and files for wallet settings & management
